+++ src/shell_builtin.rs
@@ -1,35 +1,41 @@
-use crate::command::{Command, CommandContext};
+use crate::shell::Shell;
+use std::collections::HashMap;
 use std::error::Error;
-use std::path::PathBuf;
 
-pub struct CommandChangeDirectory {}
+pub trait Command {
+    fn get_key(&self) -> Vec<&str>;
 
-impl CommandChangeDirectory {
-    pub fn new() -> Self {
-        Self {}
-    }
+    fn run(&self, ctx: CommandContext) -> Result<(), Box<dyn Error>>;
 }
 
-impl Command for CommandChangeDirectory {
-    fn get_key(&self) -> Vec<&str> {
-        vec!["Change-Directory", "cd"]
+pub struct CommandContext<'a> {
+    shell: &'a Shell,
+
+    name: String,
+    args: Vec<&'a str>,
+    envs: HashMap<String, String>,
+}
+
+impl<'a> CommandContext<'a> {
+    pub fn new(
+        shell: &'a Shell,
+        name: &str,
+        args: Vec<&'a str>,
+        envs: HashMap<String, String>,
+    ) -> Self {
+        Self {
+            shell,
+            name: name.to_string(),
+            args,
+            envs,
+        }
     }
 
-    fn run(&self, ctx: CommandContext) -> Result<(), Box<dyn Error>> {
-        let arg_0 = match ctx.get_arg(0) {
-            Some(v) => v,
-            None => return Err(Box::from("missing argument")),
-        };
+    pub fn get_shell(&self) -> &Shell {
+        self.shell
+    }
 
-        let path = PathBuf::from(arg_0);
-
-        match ctx.get_shell().update_path(path) {
-            true => {}
-            false => {
-                return Err(Box::from("invalid path"));
-            }
-        }
-
-        Ok(())
+    pub fn get_arg(&self, index: usize) -> Option<&str> {
+        self.args.get(index).cloned()
     }
 }
